+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 -   Add basic notification for create events
 -   Add statistics page
 -   Add current month event stats card
+-   Add events by employee stats card
 
 ## Fixes